requirements.txt
# Core ML dependencies
https://github.com/google-coral/pycoral/releases/download/v2.0.0/tflite_runtime-2.5.0.post1-cp39-cp39-linux_x86_64.whl
--extra-index-url https://google-coral.github.io/py-repo/
pycoral~=2.0

# Image processing
numpy>=1.16.0
Pillow>=4.0.0,<10.0

# Web application
flask==3.0.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78